CrawlingPost/requirements.txt
bs4==0.0.2
chromadb==0.4.24
langchain==0.1.16
langchain-openai==0.1.3
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
requests==2.31.0
seaborn==0.13.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
transformers==4.40.0
accelerate==0.34.2
langchain_teddynote
pymysql